Put all PyDeck selection tests into 1 file
