--- e2e_playwright/st_pydeck_chart_select_callback.py
@@ -1,29 +1,0 @@
-# Copyright (c) Streamlit Inc. (2018-2022) Snowflake Inc. (2022-2024)
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from __future__ import annotations
-
-import streamlit as st
-from shared.pydeck_utils import get_pydeck_chart
-
-
-def on_selection():
-    st.write(
-        "PyDeck selection callback:",
-        str(st.session_state.selection_callback),
-    )
-
-
-selection = get_pydeck_chart(
-    "selection_callback", selection_mode="single", on_select=on_selection
-)
